Dialog 대화 타이핑 이펙트 추가, 애니메이션 검정스크린 추가, Boss관련 기본 스크립트 추가

+++ Assets/Scripts/Character/CharacterBase.cs
@@ -20,8 +20,11 @@ public abstract class CharacterBase : MonoBehaviour {
     public float MaxExp;
     public float CurrentExp { get; set; }
 
-    public int MinimalDamage;
+    public float MinimalDamage;
     public float Damage { get => Random.Range(MinimalDamage, MinimalDamage + 3); }
+
+
+
 
     protected virtual void Awake() {
         Anim = GetComponent<Animator>();
